pyproject.toml
[project]
name = "econteacher"
version = "0.1.0"
description = "Tools to automate A-level economics teaching and tutoring"
authors = [
    {name = "jac"}
]
requires-python = "^3.12"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "pdfplumber (>=0.11.9,<0.12.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
